PvP issues p.1 same steps turn numbers

+++ server/core/views/index.php
@@ -81,35 +81,37 @@
         <div class="user-id">ID: <?php echo $_REQUEST['viewer_id'] ?></div>
     </div>
 
-    <!-- DEV -->
-    <div class="dev">
-        <div class="tests">
-            <a href="/server/index.php/user/get?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Get user</a> |
-            <a href="/server/index.php/user/add?hair_id=1&amp;face_id=1&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Add user</a> |
-            <a href="/server/index.php/user/delete?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Delete user</a> |
-            <a href="/server/index.php/user/learnStep?energy_spent=100&amp;step_id=indian_step&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Raise step level</a> |
-            <a href="/server/index.php/user/buyItem?item_id=jeans_blue&amp;color=blue&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Buy item</a> |
-            <a href="/server/index.php/user/sellItem?user_item_id=270&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Sell item</a> |
-            <a href="/server/index.php/user/equipSlot?slot_id=arms&amp;user_item_id=13&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Equip slot</a> |
-            <a href="/server/index.php/user/saveSettings?music=1&amp;sfx=0&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Save settings</a> |
-            <a href="/server/index.php/battle/test?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Send invite</a> |
-            <a href="/server/index.php/xml/load" target="_blank">Parse xmls</a> |
-            <a href="#" onclick="social.placeOrder(1);return false;">Test payments</a>
-        </div>
+    <?php if($_REQUEST['viewer_id'] == 1 || $_REQUEST['viewer_id'] == 812177) { ?>
+        <!-- DEV -->
+        <div class="dev">
+            <div class="tests">
+                <a href="/server/index.php/user/get?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Get user</a> |
+                <a href="/server/index.php/user/add?hair_id=1&amp;face_id=1&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Add user</a> |
+                <a href="/server/index.php/user/delete?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Delete user</a> |
+                <a href="/server/index.php/user/learnStep?energy_spent=100&amp;step_id=indian_step&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Raise step level</a> |
+                <a href="/server/index.php/user/buyItem?item_id=jeans_blue&amp;color=blue&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Buy item</a> |
+                <a href="/server/index.php/user/sellItem?user_item_id=270&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Sell item</a> |
+                <a href="/server/index.php/user/equipSlot?slot_id=arms&amp;user_item_id=13&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Equip slot</a> |
+                <a href="/server/index.php/user/saveSettings?music=1&amp;sfx=0&amp;viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Save settings</a> |
+                <a href="/server/index.php/battle/test?viewer_id=<?php echo $_REQUEST['viewer_id'] ?>&amp;auth_key=198a038c272bdab32fa5e0fc9a3314ef" target="_blank">Send invite</a> |
+                <a href="/server/index.php/xml/load" target="_blank">Parse xmls</a> |
+                <a href="#" onclick="social.placeOrder(1);return false;">Test payments</a>
+            </div>
 
-        <div id="buttons">
-            <input value="Start listening" onclick="pvp.channels.push(new pvp.listener('/sub?cid=<?php echo $_REQUEST['viewer_id'] ?>', pvp.onSuccess, pvp.onError));" type="button">
-            <input value="Stop listening" onclick="pvp.channels[0].stop()" type="button"><br>
-            To: <input value="<?php echo $_REQUEST['viewer_id'] ?>" name="cid" id="cid" type="text"><br>
-            Type: <input value="" name="type" id="type" type="text"><br>
-            Message: <input value="Hello" name="text" id="text" type="text">
-            <input value="Send hello" onclick="pvp.sendMessage(<?php echo $_REQUEST['viewer_id'] ?>, $('#cid').val(), $('#text').val(), $('#type').val())" type="button">
-        </div>
+            <div id="buttons">
+                <input value="Start listening" onclick="pvp.channels.push(new pvp.listener('/sub?cid=<?php echo $_REQUEST['viewer_id'] ?>', pvp.onSuccess, pvp.onError));" type="button">
+                <input value="Stop listening" onclick="pvp.channels[0].stop()" type="button"><br>
+                To: <input value="<?php echo $_REQUEST['viewer_id'] ?>" name="cid" id="cid" type="text"><br>
+                Type: <input value="" name="type" id="type" type="text"><br>
+                Message: <input value="Hello" name="text" id="text" type="text">
+                <input value="Send hello" onclick="pvp.sendMessage(<?php echo $_REQUEST['viewer_id'] ?>, $('#cid').val(), $('#text').val(), $('#type').val())" type="button">
+            </div>
 
-        <div id="messages">
+            <div id="messages">
+            </div>
         </div>
-    </div>
-    <!-- /DEV -->
+        <!-- /DEV -->
+    <?php } ?>
 
     <script type="text/javascript">
